move rotationDegree logic to parent module
clean up css classnames & properties
