+++ src/components/CreditCardRecommendation/CreditCardCarousel.tsx
@@ -1,4 +1,3 @@
-import React from 'react';
 import { clamp } from '@utils/Clamp';
 import { useThrottleTime } from '@utils/Throttle';
 import { css } from '@emotion/react';
@@ -6,9 +5,7 @@ import { gray } from '@radix-ui/colors';
 import { ArrowLeftIcon, ArrowRightIcon } from '@radix-ui/react-icons';
 import { useCreditCardRecommendationContext } from './CreditCardRecommendation';
 
-const cardWidth = 150;
-const timeOut = 500;
-const perspective = 1000;
+const throttleTime = 500;
 
 const carouselContainer = css`
   position: relative;
@@ -20,13 +17,14 @@ const carouselContainer = css`
   user-select: none;
 `;
 
-const carouselBody = css`
+const outerCylinder = css`
   width: 150px;
   aspect-ratio: 1 / 1.58;
   transform-style: preserve-3d;
+  transform: perspective(1000px) rotateY(0deg);
 `;
 
-const card = css`
+const innerCylinder = css`
   position: absolute;
   top: 0;
   right: 0;
@@ -37,11 +35,10 @@ const card = css`
   transform-origin: center;
   transform-style: preserve-3d;
   transform: rotateY(0deg);
-  transition: 500ms;
-  zindex: 0;
+  transition: ${throttleTime}ms;
 `;
 
-const cardImg = css`
+const card = css`
   position: absolute;
   top: 0;
   left: 0;
@@ -51,9 +48,12 @@ const cardImg = css`
   border-radius: 8px;
   box-shadow: 4px 4px 6px 0px rgba(0, 0, 0, 0.15);
   user-select: none;
+  cursor: pointer;
 `;
 
-const ButtonContainer = css`
+const buttonWidth = 40;
+
+const buttonContainer = css`
   position: absolute;
   display: flex;
   justify-content: space-between;
@@ -62,15 +62,15 @@ const ButtonContainer = css`
   right: 0;
   top: 0;
   bottom: 0;
+  width: calc(300px + ${buttonWidth * 2}px);
+  height: fit-content;
   margin: auto;
-  width: 400px;
-  height: calc(${cardWidth}px * 1.58);
 `;
 
-const Button = css`
-  height: 40px;
-  width: 40px;
-  border-radius: 40px;
+const button = css`
+  height: ${buttonWidth}px;
+  width: ${buttonWidth}px;
+  border-radius: ${buttonWidth}px;
   border: 0;
   display: grid;
   place-items: center;
@@ -89,10 +89,9 @@ const Button = css`
   },
 `;
 
-const snapAnimationOffset = 7;
-
 export function CreditCardCarousel() {
-  const { data, index, setIndex } = useCreditCardRecommendationContext();
+  const { data, index, rotationDegree, setIndex } =
+    useCreditCardRecommendationContext();
 
   const cards = data?.cards.map((datum) => {
     const { benefit, ...rest } = datum;
@@ -100,139 +99,23 @@ export function CreditCardCarousel() {
     return rest;
   });
 
-  const [rotationDegree, setRotationDegree] = React.useState(0);
-  const [isMouseDown, setIsMouseDown] = React.useState(false);
-  const [isMouseDragging, setIsMouseDragging] = React.useState(false);
-  const [dragDirection, setDragDirection] = React.useState<
-    'left' | 'right' | 'idle'
-  >('idle');
-  const snapAnimationId = React.useRef(0);
-  const carouselContainerRef = React.useRef<HTMLDivElement>(null);
-
-  const handleContainerMouseMove = (e: React.MouseEvent) => {
-    if (isMouseDown === true) {
-      setIsMouseDragging(true);
-
-      setRotationDegree((prev) => {
-        if (e.movementX > 0) {
-          setDragDirection('right');
-        } else if (e.movementX < 0) {
-          setDragDirection('left');
-        } else {
-          setDragDirection('idle');
-        }
-
-        const nextDegree = (prev % 360) + e.movementX;
-
-        if (nextDegree >= 90 || nextDegree <= -90) return prev;
-
-        return nextDegree;
-      });
-    }
-  };
-
-  const snapRotateCard = React.useCallback(
-    (direction: 'left' | 'right' | 'idle') => {
-      const animate = () => {
-        if (direction === 'right') {
-          setDragDirection('right');
-          setRotationDegree((prev) => {
-            const multipleOf90 = Math.floor(prev / 90) * 90;
-
-            if (
-              prev >= multipleOf90 - snapAnimationOffset &&
-              prev <= multipleOf90 + snapAnimationOffset
-            ) {
-              setDragDirection('idle');
-              cancelAnimationFrame(snapAnimationId.current);
-
-              return 0;
-            }
-
-            return (prev % 360) + snapAnimationOffset;
-          });
-        } else if (direction === 'left') {
-          setDragDirection('left');
-          setRotationDegree((prev) => {
-            const multipleOf90 = Math.floor(prev / 90) * 90;
-
-            if (
-              prev >= multipleOf90 - snapAnimationOffset &&
-              prev <= multipleOf90 + snapAnimationOffset
-            ) {
-              setDragDirection('idle');
-              cancelAnimationFrame(snapAnimationId.current);
-
-              return 0;
-            }
-
-            return (prev % 360) - snapAnimationOffset;
-          });
-        }
-
-        snapAnimationId.current = requestAnimationFrame(animate);
-      };
-
-      animate();
-
-      return () => {
-        cancelAnimationFrame(snapAnimationId.current);
-      };
-    },
-    []
-  );
-
   const showPrevCard = useThrottleTime(() => {
     if (cards !== undefined) {
       const nextCardIndex = clamp(index - 1, cards.length);
       setIndex(nextCardIndex);
     }
-  }, timeOut);
+  }, throttleTime);
 
   const showNextCard = useThrottleTime(() => {
     if (cards !== undefined) {
       const nextCardIndex = clamp(index + 1, cards.length);
       setIndex(nextCardIndex);
     }
-  }, timeOut);
-
-  const handleContainerMouseUp = () => {
-    setIsMouseDown(false);
-
-    if (isMouseDragging === true) {
-      setIsMouseDragging(false);
-      snapRotateCard(dragDirection);
-
-      if (cards !== undefined) {
-        if (dragDirection === 'left') {
-          const nextCardIndex = clamp(index + 1, cards.length);
-          setIndex(nextCardIndex);
-        } else if (dragDirection === 'right') {
-          const nextCardIndex = clamp(index - 1, cards.length);
-          setIndex(nextCardIndex);
-        }
-      }
-    }
-  };
+  }, throttleTime);
 
   return (
-    <div
-      role="presentation"
-      ref={carouselContainerRef}
-      onMouseDown={() => {
-        setIsMouseDown(true);
-      }}
-      onMouseMove={handleContainerMouseMove}
-      onMouseUp={handleContainerMouseUp}
-      onMouseLeave={handleContainerMouseUp}
-      css={carouselContainer}
-    >
-      <div
-        css={carouselBody}
-        style={{
-          transform: `perspective(${perspective}px) rotateY(${0}deg)`,
-        }}
-      >
+    <div css={carouselContainer}>
+      <div css={outerCylinder}>
         {cards !== undefined &&
           cards.map((item, i) => {
             const indexLeft = clamp(index - 1, cards.length);
@@ -241,7 +124,7 @@ export function CreditCardCarousel() {
             return (
               <span
                 key={item.id}
-                css={card}
+                css={innerCylinder}
                 style={{
                   opacity: i === index ? 1 : 0,
                   transform: `rotateY(calc(${
@@ -254,7 +137,7 @@ export function CreditCardCarousel() {
                 }}
               >
                 <img
-                  css={cardImg}
+                  css={card}
                   src={item.src}
                   alt={item.id}
                   draggable={false}
@@ -264,11 +147,11 @@ export function CreditCardCarousel() {
             );
           })}
       </div>
-      <div css={ButtonContainer}>
-        <button onClick={showPrevCard} css={Button}>
+      <div css={buttonContainer}>
+        <button onClick={showPrevCard} css={button}>
           <ArrowLeftIcon />
         </button>
-        <button onClick={showNextCard} css={Button}>
+        <button onClick={showNextCard} css={button}>
           <ArrowRightIcon />
         </button>
       </div>
