+++ src/components/CreditCardRecommendation/CreditCardRecommendation.tsx
@@ -1,23 +1,27 @@
 import React from 'react';
 import type { CardsResponse } from '@pages/Remotes';
+import { clamp } from '@utils/Clamp';
 import { CreditCardCarousel } from './CreditCardCarousel';
 import { CreditCardMeta } from './CreditCardMeta';
-// import { clamp } from './utils/clamp';
 
 interface CreditCardRecommendationProps {
   index: number;
-  // autoPlayTrigger: boolean;
+  rotationDegree: number;
   data?: CardsResponse;
   setIndex: (index: number) => void;
+  // autoPlayTrigger: boolean;
+  // setRotationDegree: (degree: number) => void;
   // setAutoPlay: React.Dispatch<React.SetStateAction<boolean>>;
 }
 
 const CreditCardRecommendationContext =
   React.createContext<CreditCardRecommendationProps>({
     index: 0,
-    // autoPlayTrigger: false,
+    rotationDegree: 0,
     data: { cards: [] },
     setIndex: () => {},
+    // setRotationDegree: () => {},
+    // autoPlayTrigger: false,
     // setAutoPlay: () => {},
   });
 
@@ -25,18 +29,117 @@ export const useCreditCardRecommendationContext = () => {
   return React.useContext(CreditCardRecommendationContext);
 };
 
+const snapAnimationOffset = 7;
+
 const CreditCardRecommendation = ({
   data,
   children,
 }: React.PropsWithChildren<Pick<CreditCardRecommendationProps, 'data'>>) => {
   const [index, setIndex] = React.useState(0);
+  const [rotationDegree, setRotationDegree] = React.useState(0);
+
+  const [isMouseDown, setIsMouseDown] = React.useState(false);
+  const [isMouseDragging, setIsMouseDragging] = React.useState(false);
+  const [dragDirection, setDragDirection] = React.useState<
+    'left' | 'right' | 'idle'
+  >('idle');
+  const snapAnimationId = React.useRef(0);
+  const carouselContainerRef = React.useRef<HTMLDivElement>(null);
+
+  const handleContainerMouseMove = (e: React.MouseEvent) => {
+    if (isMouseDown === true) {
+      setIsMouseDragging(true);
+
+      setRotationDegree((prev) => {
+        if (e.movementX > 0) {
+          setDragDirection('right');
+        } else if (e.movementX < 0) {
+          setDragDirection('left');
+        } else {
+          setDragDirection('idle');
+        }
+
+        const nextDegree = (prev % 360) + e.movementX;
+
+        if (nextDegree >= 90 || nextDegree <= -90) return prev;
+
+        return nextDegree;
+      });
+    }
+  };
+
+  const snapRotateCard = React.useCallback(
+    (direction: 'left' | 'right' | 'idle') => {
+      const animate = () => {
+        if (direction === 'right') {
+          setDragDirection('right');
+          setRotationDegree((prev) => {
+            const multipleOf90 = Math.floor(prev / 90) * 90;
+
+            if (
+              prev >= multipleOf90 - snapAnimationOffset &&
+              prev <= multipleOf90 + snapAnimationOffset
+            ) {
+              setDragDirection('idle');
+              cancelAnimationFrame(snapAnimationId.current);
+
+              return 0;
+            }
+
+            return (prev % 360) + snapAnimationOffset;
+          });
+        } else if (direction === 'left') {
+          setDragDirection('left');
+          setRotationDegree((prev) => {
+            const multipleOf90 = Math.floor(prev / 90) * 90;
+
+            if (
+              prev >= multipleOf90 - snapAnimationOffset &&
+              prev <= multipleOf90 + snapAnimationOffset
+            ) {
+              setDragDirection('idle');
+              cancelAnimationFrame(snapAnimationId.current);
+
+              return 0;
+            }
+
+            return (prev % 360) - snapAnimationOffset;
+          });
+        }
+
+        snapAnimationId.current = requestAnimationFrame(animate);
+      };
+
+      animate();
+
+      return () => {
+        cancelAnimationFrame(snapAnimationId.current);
+      };
+    },
+    []
+  );
+
+  const handleContainerMouseUp = () => {
+    setIsMouseDown(false);
+
+    if (isMouseDragging === true) {
+      setIsMouseDragging(false);
+      snapRotateCard(dragDirection);
+
+      if (data?.cards !== undefined) {
+        if (dragDirection === 'left') {
+          const nextCardIndex = clamp(index + 1, data.cards.length);
+          setIndex(nextCardIndex);
+        } else if (dragDirection === 'right') {
+          const nextCardIndex = clamp(index - 1, data.cards.length);
+          setIndex(nextCardIndex);
+        }
+      }
+    }
+  };
 
   // const autoPlayTimer = React.useRef<NodeJS.Timer>();
   // const resetAutoPlayDelay = React.useRef<NodeJS.Timeout>();
-
-  const providerValue = React.useMemo(() => {
-    return { index, setIndex, data };
-  }, [index, data]);
 
   // React.useEffect(() => {
   //   if (!autoPlay) {
@@ -71,9 +174,24 @@ const CreditCardRecommendation = ({
   //   };
   // }, [autoPlay, data?.cards, index]);
 
+  const providerValue = React.useMemo(() => {
+    return { index, rotationDegree, data, setIndex, setRotationDegree };
+  }, [index, rotationDegree, data]);
+
   return (
     <CreditCardRecommendationContext.Provider value={providerValue}>
-      {children}
+      <div
+        role="presentation"
+        ref={carouselContainerRef}
+        onMouseDown={() => {
+          setIsMouseDown(true);
+        }}
+        onMouseMove={handleContainerMouseMove}
+        onMouseUp={handleContainerMouseUp}
+        onMouseLeave={handleContainerMouseUp}
+      >
+        {children}
+      </div>
     </CreditCardRecommendationContext.Provider>
   );
 };
